word size fixups
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/trunk@34468 72102866-910b-0410-8b05-ffd578937521

+++ Source/Additions/GSMime.m
@@ -3789,8 +3789,8 @@ appendString(NSMutableData *m, NSUInteger offset, NSUInteger fold,
     }
   if (offset > fold)
     {
-      NSLog(@"Name '%@' too long for folding at %lu in header",
-	n, (unsigned long)fold);
+      NSLog(@"Name '%@' too long for folding at %"PRIuPTR" in header",
+	n, fold);
     }
 
   offset = appendBytes(md, offset, fold, ":", 1);
